+++ src/main/java/com/cMall/feedShop/review/application/service/ReviewService.java
@@ -1,6 +1,7 @@
 package com.cMall.feedShop.review.application.service;
 
-import com.cMall.feedShop.common.service.GcpStorageService;
+import com.cMall.feedShop.common.dto.UploadResult;
+import com.cMall.feedShop.common.storage.GcpStorageService;
 import com.cMall.feedShop.common.exception.BusinessException;
 import com.cMall.feedShop.common.exception.ErrorCode;
 import com.cMall.feedShop.product.domain.model.Product;
@@ -176,7 +177,7 @@ public class ReviewService {
                 // GCP Storage 서비스만 사용
                 if (gcpStorageService != null) {
                     log.info("GCP Storage 서비스 사용");
-                    List<GcpStorageService.UploadResult> uploadResults = gcpStorageService.uploadFilesWithDetails(images, "reviews");
+                    List<UploadResult> uploadResults = gcpStorageService.uploadFilesWithDetails(images, "reviews");
 
                     if (!uploadResults.isEmpty()) {
                         // UploadResult를 ReviewImage로 저장
@@ -184,7 +185,7 @@ public class ReviewService {
 
                         // URL만 추출해서 응답용으로 사용
                         imageUrls = uploadResults.stream()
-                                .map(GcpStorageService.UploadResult::getFilePath)
+                                .map(UploadResult::getFilePath)
                                 .collect(Collectors.toList());
                     }
                 } else {
@@ -349,7 +350,7 @@ public class ReviewService {
                 log.info("GCP Storage로 새 이미지 업로드 시작: reviewId={}, 이미지 수={}",
                         review.getReviewId(), newImages.size());
 
-                List<GcpStorageService.UploadResult> uploadResults =
+                List<UploadResult> uploadResults =
                         gcpStorageService.uploadFilesWithDetails(newImages, "reviews");
 
                 if (!uploadResults.isEmpty()) {
@@ -358,7 +359,7 @@ public class ReviewService {
 
                     // URL 추출
                     newImageUrls = uploadResults.stream()
-                            .map(GcpStorageService.UploadResult::getFilePath)
+                            .map(UploadResult::getFilePath)
                             .collect(Collectors.toList());
                 }
             } else {
@@ -448,13 +449,13 @@ public class ReviewService {
     }
 
     // 업로드 결과를 기존 ReviewImage 엔티티로 저장
-    private void saveReviewImagesFromUploadResults(Review review, List<GcpStorageService.UploadResult> uploadResults) {
+    private void saveReviewImagesFromUploadResults(Review review, List<UploadResult> uploadResults) {
         log.info("업로드 결과를 ReviewImage 엔티티로 저장 시작: reviewId={}, resultCount={}",
                 review.getReviewId(), uploadResults.size());
 
         try {
             for (int i = 0; i < uploadResults.size(); i++) {
-                GcpStorageService.UploadResult result = uploadResults.get(i);
+                UploadResult result = uploadResults.get(i);
 
                 ReviewImage reviewImage = ReviewImage.builder()
                         .review(review)
@@ -480,7 +481,7 @@ public class ReviewService {
 
             // 이미 업로드된 GCP Storage 파일들 삭제 (롤백)
             List<String> imageUrls = uploadResults.stream()
-                    .map(GcpStorageService.UploadResult::getFilePath)
+                    .map(UploadResult::getFilePath)
                     .collect(Collectors.toList());
             rollbackUploadedImages(imageUrls);
